PopupMenu: Add mouse movement timeout back in
This reimplements similar functionality to that removed in
d39789b021178e951b84abc5a08eac63f2d4aa10.

Unlike the previous implementation, this version will still wait for the
mouse to move over the menu before the countdown timer is allowed to
start. This should avoid the situation where menu items are accidentally
triggered on mouse-up, after a menu opens under the pointer in response
to a mouse-down.